user/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
flask-mysql-connector==1.1.0
mysql-connector-python==9.0.0
certifi==2024.7.4
requests